+++ src/components/Table/TableBody/TableBody.js
@@ -15,6 +15,7 @@ const TableBody = ({
   rowHeight,
   rowActions,
   onRowClick,
+  onCellClick,
   className,
   ...otherProps
 }) => {
@@ -54,6 +55,7 @@ const TableBody = ({
         row={ row }
         rowHeight={ rowHeight }
         onRowClick={ onRowClick }
+        onCellClick={ onCellClick }
       />
     ) })
   const isRowSelected = selectField => {
@@ -131,7 +133,9 @@ TableBody.propTypes = {
     onClick: propTypes.func,
   })),
   /** Callback fire when row click. */
-  onRowClick: propTypes.func,
+  onRowClick: propTypes.func, 
+  /** Callback fire when cell click. */
+  onCellClick: propTypes.func,
   /** For css customization. */
   className: propTypes.string,
 }